Unnecessary floats converted to ints + README.md

+++ moves.c
@@ -6,13 +6,11 @@
 /*   By: jesuserr <jesuserr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 11:47:50 by jesuserr          #+#    #+#             */
-/*   Updated: 2023/10/19 15:54:04 by jesuserr         ###   ########.fr       */
+/*   Updated: 2023/10/23 18:08:13 by jesuserr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
-
-/* Rotations and arrow keys movements */
 
 void	key_action_1(int keycode, t_fdf *fdf)
 {
@@ -40,8 +38,6 @@ void	key_action_1(int keycode, t_fdf *fdf)
 }
 
 /* Erases the image in memory; calculates and writes the new one */
-/* Keeps angles always between 0-359 degrees and prints info bar */
-/* if screen is big enough */
 
 void	action_aux(t_fdf *fdf)
 {
@@ -50,13 +46,13 @@ void	action_aux(t_fdf *fdf)
 	ft_bzero(fdf->img.addr, WIDTH * HEIGHT * fdf->img.bpp / 8);
 	projection(fdf);
 	mlx_put_image_to_window(fdf->mlx, fdf->mlx_win, fdf->img.img, 0, 0);
-	str = ft_itoa((int)fdf->player.x_pos);
+	str = ft_itoa(fdf->player.x_pos);
 	mlx_string_put(fdf->mlx, fdf->mlx_win, 500, 10, WHITE, str);
 	free(str);
-	str = ft_itoa((int)fdf->player.y_pos);
+	str = ft_itoa(fdf->player.y_pos);
 	mlx_string_put(fdf->mlx, fdf->mlx_win, 500, 30, WHITE, str);
 	free(str);
-	str = ft_itoa((int)fdf->player.angle);
+	str = ft_itoa(fdf->player.angle);
 	mlx_string_put(fdf->mlx, fdf->mlx_win, 500, 50, WHITE, str);
 	free(str);
 }
